Minor integration test cleanup (#435)

+++ test/integration/sifchain_logs.sh
@@ -1,9 +1,15 @@
 #!/bin/bash
 
-set -e
+set -xe
 
 basedir=$(dirname $0)
 . $basedir/vagrantenv.sh
+
+if [ -z $EBRELAYER_LOG ]
+then
+  echo $0: must specify EBRELAYER_LOG
+  exit 1
+fi
 
 output=$datadir/sifchaintxs
 rm -f $output*
@@ -12,5 +18,5 @@ hashes=$(cat $EBRELAYER_LOG | grep "^txhash: " | sed -e "s/txhash: //")
 for i in $hashes
 do
   sifnodecli q tx $i >> $output.txt
-  sifnodecli q tx $i -o json | jq -c >> $output.json
+  sifnodecli q tx $i -o json | jq -c . >> $output.json
 done
